refactor(reducer): Split reducer logic
Split to book logic and cart logic.
Fix changes in CartTable and BookList.
